This update enables the (un)deployment of service bags (zips) through servicesadmin.
git-svn-id: https://svn.duraspace.org/duracloud-priv@200 3c0c4179-e91d-42e7-a082-8e2792ce3676

+++ trunk/services/servicesutil/src/test/java/org/duracloud/servicesutil/osgi/ServiceInstallTestBase.java
@@ -1,4 +1,3 @@
-
 package org.duracloud.servicesutil.osgi;
 
 import java.io.BufferedReader;
@@ -9,42 +8,107 @@ import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
+import java.util.zip.ZipFile;
+import java.util.zip.ZipEntry;
+import java.util.Enumeration;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.junit.Assert;
+import static org.junit.Assert.assertTrue;
 
 import static junit.framework.Assert.assertEquals;
 
+/**
+ * @author Andrew Woods
+ * Date: Oct 1, 2009
+ */
 public class ServiceInstallTestBase {
 
     private final Logger log = LoggerFactory.getLogger(getClass());
 
-    protected final static String TEST_BUNDLE_FILE_NAME = "junk-bundle.jar";
+    private final static String BASE_DIR_PROP = "base.dir";
 
-    protected final static String BUNDLE_TEXT =
+    protected final static String DUMMY_BUNDLE_FILE_NAME = "junk-bundle.jar";
+
+    protected final static String DUMMY_BUNDLE_TEXT =
             "normally-bundle-is-a-jar-not-text";
 
-    protected InputStream getTestBundle() throws FileNotFoundException {
-        return new ByteArrayInputStream(BUNDLE_TEXT.getBytes());
+    protected final static String ZIP_BAG_FILE_NAME = "replicationservice-1.0.0.zip";
+
+    protected final static String sep = File.separator;
+
+    protected InputStream getDummyBundle() throws FileNotFoundException {
+        return new ByteArrayInputStream(DUMMY_BUNDLE_TEXT.getBytes());
     }
 
-    protected void verifyTestBundleInstalled(String home, boolean exists)
+    protected InputStream getZipBag() throws FileNotFoundException {
+        String baseDir = System.getProperty(BASE_DIR_PROP);
+        Assert.assertNotNull(baseDir);
+
+        String resourceDir = baseDir + sep + "src/test/resources/";
+        File zipBagFile = new File(resourceDir + ZIP_BAG_FILE_NAME);
+
+        return new FileInputStream(zipBagFile);
+    }
+
+    protected void verifyDummyBundleInstalled(String home, boolean exists)
             throws FileNotFoundException, IOException {
-        File file = new File(home + File.separator + TEST_BUNDLE_FILE_NAME);
+        File file = new File(home + sep + DUMMY_BUNDLE_FILE_NAME);
         assertEquals(exists, file.exists());
 
         if (exists) {
             BufferedReader br =
                     new BufferedReader(new InputStreamReader(new FileInputStream(file)));
             String line = br.readLine();
-            assertEquals(BUNDLE_TEXT, line);
+            assertEquals(DUMMY_BUNDLE_TEXT, line);
             br.close();
         }
     }
 
-    protected void deleteTestBundle(String home) {
-        File file = new File(home + File.separator + TEST_BUNDLE_FILE_NAME);
-        file.delete();
+    protected void verifyZipBagInstalled(String home, boolean exists)
+            throws FileNotFoundException, IOException {
+        File file = new File(getAttic(home) + sep + ZIP_BAG_FILE_NAME);
+        assertEquals("file: " + file.getAbsoluteFile(), exists, file.exists());
+
+        if (exists) {
+            ZipFile zip = new ZipFile(file);
+            assertTrue(zip.size() > 0);
+
+            Enumeration entries = zip.entries();
+            while (entries.hasMoreElements()) {
+                ZipEntry entry = (ZipEntry) entries.nextElement();
+                File entryBundle = new File(home + sep + entry.getName());
+                assertTrue(entryBundle.exists());
+            }
+        }
     }
 
+    protected void deleteDummyBundle(String home) {
+        File file = new File(home + sep + DUMMY_BUNDLE_FILE_NAME);
+        file.delete();
+
+        File atticFile = new File(getAttic(home) + sep + DUMMY_BUNDLE_FILE_NAME);
+        atticFile.delete();
+    }
+
+    protected void deleteZipBagBundles(String home) throws IOException {
+        File file = new File(getAttic(home) + sep + ZIP_BAG_FILE_NAME);
+
+        if (file.exists()) {
+            ZipFile zip = new ZipFile(file);
+
+            Enumeration entries = zip.entries();
+            while (entries.hasMoreElements()) {
+                ZipEntry entry = (ZipEntry) entries.nextElement();
+                File entryBundle = new File(home + sep + entry.getName());
+                entryBundle.delete();
+            }
+            file.delete();
+        }
+    }
+
+    protected File getAttic(String home) {
+        return new File(home + sep + "attic");
+    }
 }
